fix(create): createコンポーネントからeditor,tag-formを分割
