+++ src/app/notes/create/create.component.html
@@ -18,43 +18,14 @@
           >タイトルは255文字以内です</mat-error
         >
       </mat-form-field>
-      <mat-form-field class="editor__tags">
-        <mat-chip-list #chipList aria-label="タグセレクション">
-          <mat-chip
-            *ngFor="let tag of tags"
-            [selectable]="selectable"
-            [removable]="removable"
-            (removed)="remove(tag)"
-          >
-            {{ tag }}
-            <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
-          </mat-chip>
-          <input
-            placeholder="タグ"
-            #tagInput
-            formControlName="tag"
-            [matAutocomplete]="auto"
-            [matChipInputFor]="chipList"
-            [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
-            (matChipInputTokenEnd)="add($event)"
-          />
-        </mat-chip-list>
-        <mat-autocomplete
-          #auto="matAutocomplete"
-          (optionSelected)="selected($event)"
-        >
-          <mat-option *ngFor="let tag of allTags" [value]="tag.value">
-            {{ tag.value }}({{ tag.count }})
-          </mat-option>
-        </mat-autocomplete>
-        <mat-hint align="end">{{ tags?.length || 0 }}/10</mat-hint>
-        <mat-hint *ngIf="isTagWordOver" class="tag-error"
-          >タグは50文字以内です</mat-hint
-        >
-      </mat-form-field>
+      <app-tag-form
+        class="editor__tags"
+        [parentForm]="form"
+        [tags]="tags"
+      ></app-tag-form>
     </div>
     <div class="editor__content">
-      <div [froalaEditor]="options" formControlName="editorContent"></div>
+      <app-editor [parentForm]="form"></app-editor>
     </div>
     <div class="editor__buttons">
       <mat-slide-toggle class="public-option" formControlName="isPublic"
