+++ src/app/notes/editor/editor.component.html
@@ -1,0 +1,3 @@
+<div [formGroup]="parentForm">
+  <div [froalaEditor]="options" formControlName="editorContent"></div>
+</div>
