+++ src/app/notes/tag-form/tag-form.component.html
@@ -1,0 +1,36 @@
+<div [formGroup]="parentForm">
+  <mat-form-field class="tag-form">
+    <mat-chip-list #chipList aria-label="タグセレクション">
+      <mat-chip
+        *ngFor="let tag of tags"
+        [selectable]="selectable"
+        [removable]="removable"
+        (removed)="remove(tag)"
+      >
+        {{ tag }}
+        <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
+      </mat-chip>
+      <input
+        placeholder="タグ"
+        #tagInput
+        formControlName="tag"
+        [matAutocomplete]="auto"
+        [matChipInputFor]="chipList"
+        [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
+        (matChipInputTokenEnd)="add($event)"
+      />
+    </mat-chip-list>
+    <mat-autocomplete
+      #auto="matAutocomplete"
+      (optionSelected)="selected($event)"
+    >
+      <mat-option *ngFor="let tag of allTags" [value]="tag.value">
+        {{ tag.value }}({{ tag.count }})
+      </mat-option>
+    </mat-autocomplete>
+    <mat-hint align="end">{{ tags?.length || 0 }}/10</mat-hint>
+    <mat-hint *ngIf="isTagWordOver" class="tag-error"
+      >タグは50文字以内です</mat-hint
+    >
+  </mat-form-field>
+</div>
