+++ src/app/notes/tag-form/tag-form.component.scss
@@ -1,0 +1,6 @@
+.tag-form {
+  width: 100%;
+}
+.tag-error {
+  color: #f44336;
+}
